requirements.txt
fastapi
uvicorn
mysql-connector-python
platform
hashlib